Move results command to its forever home
