+++ planet/cli/subscriptions.py
@@ -10,13 +10,14 @@ from planet.cli.cmds import coro, translate_exceptions
 from planet.cli.io import echo_json
 from planet.exceptions import PlanetError
 
-# A collection of fake subscriptions for testing. Tests will
-# monkeypatch this attribute.
+# Collections of fake subscriptions and results for testing. Tests will
+# monkeypatch these attributes.
 _fake_subs = None
 _fake_sub_results = None
 
 
-# For test use.
+# The following 5 functions are for use in CLI commands until there
+# is subscriptions Python client to call on.
 def _count_fake_subs():
     return len(_fake_subs)
 
@@ -233,3 +234,62 @@ async def describe_subscription(ctx, subscription_id, pretty):
     # doesn't exist yet).
 
     echo_json(sub, pretty)
+
+
+@subscriptions.command(name='results')
+@click.argument('subscription_id')
+@click.option('--pretty', is_flag=True, help='Pretty-print output.')
+@click.option(
+    '--status',
+    type=click.Choice(["created", "queued", "processing", "failed",
+                       "success"]),
+    multiple=True,
+    default=None,
+    help="Select subscription results in one or more states. Default: all.")
+@click.option('--limit',
+              type=int,
+              default=100,
+              help='Maximum number of results to return. Defaults to 100.')
+# TODO: the following 3 options.
+# –created: timestamp instant or range.
+# –updated: timestamp instant or range.
+# –completed: timestamp instant or range.
+@click.pass_context
+@translate_exceptions
+@coro
+async def list_subscription_results(ctx,
+                                    subscription_id,
+                                    pretty,
+                                    status,
+                                    limit):
+    """Gets results of a subscription and prints the API response.
+
+    This implementation is only a placeholder. To begin, instead
+    of mocking calls to the Subscriptions API, we'll use a
+    collection of fake subscriptions (the all_subs object).
+    After we refactor we will change to mocking the API.
+
+    """
+    # Begin fake subscriptions service. Note that the Subscriptions
+    # API will report missing keys differently, but the Python API
+    # *will* raise PlanetError like this.
+
+    try:
+        # Filter by status, like the Subscriptions API does.
+        if status:
+            select_results = (res
+                              for res in _get_fake_sub_results(subscription_id)
+                              if res['status'] in status)
+        else:
+            select_results = _get_fake_sub_results(subscription_id)
+
+        filtered_results = itertools.islice(select_results, limit)
+    except KeyError:
+        raise PlanetError(f"No such subscription: {subscription_id!r}")
+
+    # End fake subscriptions service. After we refactor we will get
+    # the "sub" from a method in planet.clients.subscriptions (which
+    # doesn't exist yet).
+
+    for result in filtered_results:
+        echo_json(result, pretty)
